Refactor build process to cater for async steps
This fixes #9. Before, there was a chance of a race condition
where gitbook would copy files before they'd been built.

The init hook will wait for promises, so we just return promises
from the build steps. To optimize things, we should consider
building the JS ahead of time, as opposed to when generating the
book, since it doesn't have runtime dependencies like the css does.