Replace logo reference in README with img in template
This will only work for logos added with markdown embedding, and
is a bit of a hack, honestly. We may want to do this after the
markdown has been parsed and rendered, by using cheerio or some
other dom processing library.

The logo is added in the build, and added to the css as well for
the use of the logo in the summary – this fixes #15. The first H1
heading of the readme will be used as the main title of the project.